+++ packages/op/tests/types/di.test.ts
@@ -1,7 +1,7 @@
 import { describe, expectTypeOf, test } from "vitest";
 import { Op } from "../../src/index.js";
 import { type Blocking, type EmptyMeta, type InferOpMeta } from "../../src/internal/index.js";
-import { type IsRunnable } from "../../src/core/types.js";
+import { type IsRunnable } from "../../src/core/meta.js";
 import { DI, type Dependency } from "../../src/di/index.js";
 import type {
   Deps,